chore(webrtc): WIP refactor webrtc - cleanup
